Refactor filter component to lift state from the parent component
